add task+ get token from login form + static up
