--- .~lock.Documentação_TO_DO.odt#
@@ -1,1 +1,0 @@
-,vagner,nurnberg-HP-G42-Notebook-PC,19.03.2020 14:46,file:///home/vagner/.config/libreoffice/4;
