+++ projeto/frontend/style.css
@@ -48,6 +48,7 @@ span.psw {
 }
 
 
+
 @media screen and (max-width: 300px) {
   span.psw {
      display: block;
