+++ projeto/frontend/tarefas/todos.html
@@ -8,14 +8,15 @@
         <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
         <link href="todos.css" rel="stylesheet"/>
         <link href="../style.css" rel="stylesheet" />
+        <script src="crud_tarefas.js"></script>
         <script src="aut_tarefas.js"></script>    
     </head>
     
     <body>
         <div id="myDIV" class="header">
             <h2>My To Do List</h2>
-            <input type="text" id="myInput" placeholder="Title...">
-            <span onclick="newElement()" class="addBtn">Add</span>
+            <input id="texto_de_tarefa" type="text" placeholder="Entre com a sua tarefa" name="uname" required>
+            <input id="botao_de_adicionar" type="button" onclick="createTask()" class="btn btn-success" value="Adicionar tarefa" ></input>
           </div>    
           
           <ul id="concluidas">
