--- projeto/js_api/tarefas/todos.css
@@ -1,12 +1,0 @@
-body{
-    background-color:darkslategray;
-    color:white;
-}
-
-ul{
-    list-style:none;
-}
-
-header{
-    text-align: center;
-}
